Moved ZshConstants class to a utils package
